In the future, many special functions whose run-time scales maybe poorly with the target error may come to be added.  These will all need to take a target error and optionally return an estimate of the achieved error (which can sometimes be much better than what was requested).
Since many but not all callers may not care to either alter error target
requests or receive error estimate, we want a signature with optionality/
defaults.  For the returned error estimate, this means a `ptr` type since
a default `nil` can denote "don't want it" while a `var` would require more
work on the many callers for the sake of the few.

For `hoeffding` in particular function, the calculation is to machine
precision for the generic floating type `T`.  { `est` fill in could be
better about that, as per the new comment. }

Also, rename `T` -> `F` since the generic type has to be float-ish (`exp`
will never make much sense to define on `int`s and even if it did, `t`
is notionally a floating point (or at least rational) type anyway.